Extract configuration management
Adds a command exposing start_active_view
detect_supportable_view now subscribes seperately to open/activate events
