+++ __init__.py
@@ -2,3 +2,4 @@ from .core.main import *
 from .core.panels import *
 from .features.completion import *
 from .features.diagnostics import *
+from .features.configuration import *
