+++ core/main.py
@@ -1,6 +1,5 @@
 import os
 import subprocess
-import webbrowser
 
 try:
     from typing import Any, List, Dict, Tuple, Callable, Optional
@@ -18,14 +17,13 @@ from .protocol import (
     Request, Notification, Point, Range, SymbolKind
 )
 from .settings import (
-    ClientConfig, settings, client_configs, load_settings, unload_settings, PLUGIN_NAME
+    ClientConfig, settings, load_settings, unload_settings, PLUGIN_NAME
 )
 from .logging import debug, exception_log, server_log
 from .rpc import Client
-from .workspace import get_project_path, enable_in_project, disable_in_project
+from .workspace import get_project_path
 from .configurations import (
-    config_for_scope, is_supported_view, is_supported_syntax, is_supportable_syntax, get_default_client_config,
-    clear_window_client_configs, get_scope_client_config
+    config_for_scope, is_supported_view, is_supported_syntax, is_supportable_syntax
 )
 from .clients import (
     client_for_view, add_window_client, window_clients, check_window_unloaded, unload_old_clients,
@@ -107,136 +105,6 @@ def initialize_on_open(view: sublime.View):
                 get_window_client(view, window, config)
         else:
             debug(config.name, 'is not enabled')
-    else:
-        available_config = get_default_client_config(view)
-        if available_config:
-            show_enable_config(view, available_config)
-
-
-def extract_syntax_name(syntax_file: str) -> str:
-    return syntax_file.split('/')[-1].split('.')[0]
-
-
-def show_enable_config(view: sublime.View, config: ClientConfig):
-    syntax = str(view.settings().get("syntax", ""))
-    message = "LSP has found a language server for {}. Run \"Setup Language Server\" to start using it".format(
-        extract_syntax_name(syntax)
-    )
-    window = view.window()
-    if window:
-        window.status_message(message)
-
-
-class LspEnableLanguageServerGloballyCommand(sublime_plugin.WindowCommand):
-    def run(self):
-        view = self.window.active_view()
-        available_config = get_scope_client_config(view, client_configs.defaults) or get_default_client_config(view)
-        if available_config:
-            client_configs.enable(available_config.name)
-            clear_window_client_configs(self.window)
-            sublime.set_timeout_async(start_active_view, 500)
-            self.window.status_message("{} enabled, starting server...".format(available_config.name))
-            return
-
-        self.window.status_message("No config available to enable")
-
-
-class LspEnableLanguageServerInProjectCommand(sublime_plugin.WindowCommand):
-    def run(self):
-        view = self.window.active_view()
-
-        # if no default_config, nothing we can do.
-        default_config = get_default_client_config(view)
-        if default_config:
-            enable_in_project(self.window, default_config.name)
-            clear_window_client_configs(self.window)
-            sublime.set_timeout_async(start_active_view, 500)
-            self.window.status_message("{} enabled in project, starting server...".format(default_config.name))
-        else:
-            self.window.status_message("No config available to enable")
-
-
-class LspDisableLanguageServerGloballyCommand(sublime_plugin.WindowCommand):
-    def run(self):
-        view = self.window.active_view()
-        global_config = get_scope_client_config(view, client_configs.all)
-        if global_config:
-            client_configs.disable(global_config.name)
-            clear_window_client_configs(self.window)
-            sublime.set_timeout_async(lambda: unload_window_clients(self.window.id()), 500)
-            self.window.status_message("{} disabled, shutting down server...".format(global_config.name))
-            return
-
-        self.window.status_message("No config available to disable")
-
-
-class LspDisableLanguageServerInProjectCommand(sublime_plugin.WindowCommand):
-    def run(self):
-        view = self.window.active_view()
-        global_config = get_scope_client_config(view, client_configs.defaults)
-        if global_config:
-            disable_in_project(self.window, global_config.name)
-            clear_window_client_configs(self.window)
-            sublime.set_timeout_async(lambda: unload_window_clients(self.window.id()), 500)
-            self.window.status_message("{} disabled in project, shutting down server...".format(global_config.name))
-            return
-        else:
-            self.window.status_message("No config available to disable")
-
-
-supported_syntax_template = '''
-Installation steps:
-
-* Open the [LSP documentation](https://lsp.readthedocs.io)
-* Read the instructions for {}
-* Install the language server on your system
-* Choose an option below to start the server
-
-Enable: [Globally](#enable_globally) | [This Project Only](#enable_project)
-'''
-
-unsupported_syntax_template = """
-*LSP has no built-in configuration for a {} language server*
-
-Visit [langserver.org](https://langserver.org) to find out if a language server exists for this language."""
-
-
-setup_css = ".mdpopups .lsp_documentation { margin: 20px; font-family: sans-serif; font-size: 1.2rem; line-height: 2}"
-
-
-class LspSetupLanguageServerCommand(sublime_plugin.WindowCommand):
-    def run(self):
-        view = self.window.active_view()
-        syntax = view.settings().get("syntax")
-        available_config = get_default_client_config(view)
-
-        syntax_name = extract_syntax_name(syntax)
-        title = "# Language Server for {}\n".format(syntax_name)
-
-        if available_config:
-            content = supported_syntax_template.format(syntax_name)
-        else:
-            title = "# No Language Server support"
-            content = unsupported_syntax_template.format(syntax_name)
-
-        mdpopups.show_popup(
-            view,
-            "\n".join([title, content]),
-            css=setup_css,
-            md=True,
-            wrapper_class="lsp_documentation",
-            max_width=800,
-            max_height=600,
-            on_navigate=self.on_hover_navigate
-        )
-
-    def on_hover_navigate(self, href):
-        if href == "#enable_globally":
-            self.window.run_command("lsp_enable_language_server_globally")
-        elif href == "#enable_project":
-            self.window.run_command("lsp_enable_language_server_in_project")
-        else:
-            webbrowser.open_new_tab(href)
 
 
 def handle_initialize_result(result, client, window, config):
@@ -963,6 +831,14 @@ class LspRestartClientCommand(sublime_plugin.TextCommand):
     def run(self, edit):
         window = self.view.window()
         unload_window_clients(window.id())
+
+
+class LspStartClientCommand(sublime_plugin.TextCommand):
+    def is_enabled(self):
+        return is_supported_view(self.view)
+
+    def run(self, edit):
+        start_active_view()
 
 
 class LspApplyWorkspaceEditCommand(sublime_plugin.WindowCommand):
